LayoutTests/platform/chromium/http/tests/inspector-enabled/console-exception-while-no-inspector-expected.txt

CONSOLE MESSAGE: line 6: Uncaught 2013
Tests that console will NOT contain stack trace for exception thrown when inspector front-end was closed. Bug 109427.

SUCCESS: message doesn't have stack trace
TEST COMPLETE.
